+++ CHANGELOG.md
@@ -5,6 +5,7 @@
  - [#88](https://github.com/projectapheleia/avl/issues/88) Coverbin(): getstddev asserts if _count_ <=1
  - [#89](https://github.com/projectapheleia/avl/issues/89) IndexedScoreboard(): Ordering of checks can be improved by threading the before and after ports
  - [#90](https://github.com/projectapheleia/avl/issues/90) Object(): manipulation of _auto_random_ can cause false constraint fails
+ - [#85](https://github.com/projectapheleia/avl/issues/85) Imports Steps and TimeUnitWithoutSteps from cocotb.simtime will break with cocotb 2.1
 
 ## [v0.5.1] - 2026-03-05